Reorganized source tree to get rid of extraneous folders.
